# wallbox/wallbox_mobile/API-DOC/swagger2.yml
openapi: 3.0.1
info:
  title: Wallbox Project API
  description: API for Wallbox mobile applications (Odoo v18 backend)
  version: "1.1.0"
  contact:
    name: API Support
    email: munafkhan732@gmail.com
servers:
  - url: https://qariqa.qariqa.com/v1
    description: Production server
  - url: https://qariqa-staging.qariqa.com/v1
    description: Staging server
...
paths:
...
  /rfid-tags:
    get:
      tags: [RFID Tags]
      summary: List RFID Tags
      description: Returns a list of RFID tags for stations owned by the user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFIDTagListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [RFID Tags]
      summary: Create RFID Tag
      description: Creates a new RFID tag and associates it with one of the user's charging stations.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RFIDTagCreateRequest'
      responses:
        '201':
          description: RFID tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFIDTagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /rfid-tags/{tag_id}:
    parameters:
      - $ref: '#/components/parameters/TagId'
    get:
      tags: [RFID Tags]
      summary: Get RFID Tag Details
      description: Returns detailed information for a specific RFID tag.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFIDTagResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags: [RFID Tags]
      summary: Update RFID Tag
      description: Updates information for a specific RFID tag.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RFIDTagUpdateRequest'
      responses:
        '200':
          description: RFID tag updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFIDTagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [RFID Tags]
      summary: Delete RFID Tag
      description: Deletes an RFID tag.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: RFID tag deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # Charging Requests
...
components:
  schemas:
...
    RFIDTag:
      type: object
      properties:
        id:
          type: integer
        tag_id:
          type: string
        is_allowed:
          type: boolean
        charging_station_id:
          type: integer
        charging_station_name:
          type: string
    RFIDTagCreateRequest:
      type: object
      properties:
        tag_id:
          type: string
          description: "The unique identifier of the RFID tag"
        charging_station_id:
          type: integer
          description: "The ID of the charging station to associate this tag with"
        is_allowed:
          type: boolean
          description: "Whether the tag is currently active"
          default: true
      required: [tag_id, charging_station_id]
    RFIDTagUpdateRequest:
      type: object
      properties:
        tag_id:
          type: string
        charging_station_id:
          type: integer
        is_allowed:
          type: boolean
    RFIDTagResponse:
      type: object
      properties:
        status:
          type: boolean
        message:
          type: string
        rfid_tag:
          $ref: '#/components/schemas/RFIDTag'
    RFIDTagListResponse:
      type: object
      properties:
        status:
          type: boolean
        rfid_tags:
          type: array
          items:
            $ref: '#/components/schemas/RFIDTag'
        count:
          type: integer

    # Charging Request Schemas
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation was successful"
    ErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              nullable: true
      required: [status, error]
  parameters:
...
    TagId:
      name: tag_id
      in: path
      required: true
      schema:
        type: integer
      description: RFID Tag ID
...
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "VALIDATION_ERROR"
              message: "Required fields are missing"
    Unauthorized:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "TOKEN_INVALID"
              message: "Invalid or expired authentication token"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "The requested resource was not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT